add support for not available commands and empty log files
When a command is wrapped with "support_cmd", the availability of the
command is checked with "which" and a suitable error message is printed
if the command is not found.

If a log file created with "support_log_file" is empty, it is omitted
from the archive and an error message is printed in support-pack.txt.
This is so that a support-pack config that looks like:

    support_cmd <cmd1> | support_log_file <file1>
    support_cmd <cmd2> | support_log_file <file2>
    support_cmd <cmd3> | support_log_file <file3>

does not contain lots of empty files.

"nmcli" and "cukinia" can now safely be added to default.conf now that
unavailable commands are gracefully supported.

Change-Id: I6439a8b22d8e08fca3370592bf751128a3d83ef7